pyproject.toml
[project]
name = "auto-qa-pipeline"
version = "0.1.0"
description = "Automatic Qualitative Analysis Pipeline with Human-in-the-Loop"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0"
]
